feat(daemon): split loaders
