+++ app/daemon/src/lib.rs
@@ -6,7 +6,6 @@ pub mod config;
 pub mod feature;
 pub mod gateway;
 pub mod loader;
-pub mod utils;
 
 pub use channel::router::{ChannelRouter, RoutingRule};
 pub use config::DaemonConfig;
