--- app/daemon/src/utils.rs
@@ -1,29 +1,0 @@
-//! Gateway utility functions.
-
-/// Expand `${VAR}` patterns in a string with environment variable values.
-///
-/// Unknown variables are replaced with an empty string.
-pub fn expand_env_vars(input: &str) -> String {
-    let mut result = String::with_capacity(input.len());
-    let mut chars = input.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        if ch == '$' && chars.peek() == Some(&'{') {
-            chars.next(); // consume '{'
-            let mut var_name = String::new();
-            for c in chars.by_ref() {
-                if c == '}' {
-                    break;
-                }
-                var_name.push(c);
-            }
-            if let Ok(val) = std::env::var(&var_name) {
-                result.push_str(&val);
-            }
-        } else {
-            result.push(ch);
-        }
-    }
-
-    result
-}
